Made a separation of the functionality between a core and a plugin
