+++ stepik-union/src/main/java/org/stepik/plugin/refactoring/safeDelete/StepikSafeDeleteProcessorDelegate.java
@@ -13,7 +13,7 @@ import org.jetbrains.annotations.Nullable;
 import java.util.Collection;
 import java.util.List;
 
-import static org.stepik.plugin.utils.ProjectFilesUtils.isNotMovableOrRenameElement;
+import static org.stepik.plugin.utils.ProjectPsiFilesUtils.isNotMovableOrRenameElement;
 
 /**
  * @author meanmail
