+++ stepik-union/src/main/java/org/stepik/plugin/utils/PresentationDataUtils.java
@@ -17,28 +17,20 @@ import com.jetbrains.tmp.learning.courseFormat.StudyStatus;
 import com.jetbrains.tmp.learning.courseFormat.Task;
 import icons.InteractiveLearningIcons;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
+import org.stepik.core.utils.PresentationUtils;
+import org.stepik.core.utils.ProjectFilesUtils;
 
 import javax.swing.*;
-import java.awt.*;
 import java.util.HashMap;
 
-import static org.stepik.plugin.utils.ProjectFilesUtils.getParent;
-import static org.stepik.plugin.utils.ProjectFilesUtils.getRelativePath;
-import static org.stepik.plugin.utils.ProjectFilesUtils.isHideDir;
-import static org.stepik.plugin.utils.ProjectFilesUtils.isSandbox;
-import static org.stepik.plugin.utils.ProjectFilesUtils.isStudyItemDir;
-import static org.stepik.plugin.utils.ProjectFilesUtils.isTaskHtmlFile;
-import static org.stepik.plugin.utils.ProjectFilesUtils.isUtilDir;
-import static org.stepik.plugin.utils.ProjectFilesUtils.isWithinHideDir;
-import static org.stepik.plugin.utils.ProjectFilesUtils.isWithinSandbox;
-import static org.stepik.plugin.utils.ProjectFilesUtils.isWithinSrc;
-import static org.stepik.plugin.utils.ProjectFilesUtils.isWithinUtil;
+import static org.stepik.core.utils.PresentationUtils.getColor;
+import static org.stepik.core.utils.PresentationUtils.getIconMap;
+import static org.stepik.plugin.utils.ProjectPsiFilesUtils.getRelativePath;
 
 /**
  * @author meanmail
  */
-public class PresentationUtils {
+public class PresentationDataUtils {
 
     public static void updatePresentationData(@NotNull PresentationData data, @NotNull PsiDirectory psiDirectory) {
         Project project = psiDirectory.getProject();
@@ -86,62 +78,6 @@ public class PresentationUtils {
             data.setPresentableText(valueName);
     }
 
-    private static final HashMap<Object, HashMap<StudyStatus, Icon>> iconMap = new HashMap<>();
-
-    @Nullable
-    private static HashMap<StudyStatus, Icon> getIconMap(@Nullable Object subject) {
-        HashMap<StudyStatus, Icon> result = iconMap.get(subject);
-        if (result != null)
-            return result;
-
-        if (subject instanceof Course) {
-            HashMap<StudyStatus, Icon> map = new HashMap<>();
-            map.put(StudyStatus.Unchecked, InteractiveLearningIcons.Course);
-            map.put(StudyStatus.Solved, InteractiveLearningIcons.CourseCompl);
-            map.put(StudyStatus.Failed, InteractiveLearningIcons.Course);
-            iconMap.put(subject, map);
-            return map;
-        }
-        if (subject instanceof Section) {
-            HashMap<StudyStatus, Icon> map = new HashMap<>();
-            map.put(StudyStatus.Unchecked, InteractiveLearningIcons.Section);
-            map.put(StudyStatus.Solved, InteractiveLearningIcons.SectionCompl);
-            map.put(StudyStatus.Failed, InteractiveLearningIcons.Section);
-            iconMap.put(subject, map);
-            return map;
-        }
-        if (subject instanceof Lesson) {
-            HashMap<StudyStatus, Icon> map = new HashMap<>();
-            map.put(StudyStatus.Unchecked, InteractiveLearningIcons.Lesson);
-            map.put(StudyStatus.Solved, InteractiveLearningIcons.LessonCompl);
-            map.put(StudyStatus.Failed, InteractiveLearningIcons.Lesson);
-            iconMap.put(subject, map);
-            return map;
-        }
-        if (subject instanceof Task) {
-            HashMap<StudyStatus, Icon> map = new HashMap<>();
-            map.put(StudyStatus.Unchecked, InteractiveLearningIcons.Task);
-            map.put(StudyStatus.Solved, InteractiveLearningIcons.TaskCompl);
-            map.put(StudyStatus.Failed, InteractiveLearningIcons.TaskProbl);
-            iconMap.put(subject, map);
-            return map;
-        }
-        return null;
-    }
-
-    @NotNull
-    private static JBColor getColor(@NotNull StudyStatus status) {
-        switch (status) {
-            case Unchecked:
-                return JBColor.BLACK;
-            case Solved:
-                return new JBColor(new Color(0, 134, 0), new Color(98, 150, 85));
-            case Failed:
-                return JBColor.RED;
-        }
-        return JBColor.BLACK;
-    }
-
     private static void setAttributes(@NotNull PresentationData data, @NotNull StudyItem item) {
         String text = item.getName();
         HashMap<StudyStatus, Icon> iconMap = getIconMap(item);
@@ -165,34 +101,13 @@ public class PresentationUtils {
             return false;
         }
         String path = psiDirectory.getVirtualFile().getPath();
-        String relPath = getRelativePath(basePath, path);
+        String relPath = ProjectFilesUtils.getRelativePath(basePath, path);
 
-        return isVisibleDirectory(relPath);
-    }
-
-    public static boolean isVisibleDirectory(@NotNull String relPath) {
-        if (isHideDir(relPath) || isWithinHideDir(relPath)) {
-            return false;
-        }
-
-        if (isSandbox(relPath) || isStudyItemDir(relPath) || isUtilDir(relPath)) {
-            return true;
-        }
-
-        return isWithinSrc(relPath) || isWithinSandbox(relPath) || isWithinUtil(relPath);
+        return PresentationUtils.isVisibleDirectory(relPath);
     }
 
     public static boolean isVisibleFile(@NotNull PsiFile psiFile) {
         String path = getRelativePath(psiFile);
-        return isVisibleFile(path);
-    }
-
-    public static boolean isVisibleFile(@NotNull String relFilePath) {
-        String parentDir = getParent(relFilePath);
-        if (parentDir == null || isTaskHtmlFile(relFilePath) || !isVisibleDirectory(parentDir)) {
-            return false;
-        }
-
-        return isWithinSrc(relFilePath) || isWithinSandbox(relFilePath) || isWithinUtil(relFilePath);
+        return PresentationUtils.isVisibleFile(path);
     }
 }
